Moving getLessons to separate lessons file
